# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "liber API",
    "description": "",
    "license": {
      "name": "GPL-3.0-or-later",
      "identifier": "GPL-3.0-or-later"
    },
    "version": "2"
  },
  "paths": {
    "/api/v2/bookmarks": {
      "get": {
        "tags": [
          "crate::bookmarks"
        ],
        "operationId": "list_bookmarks",
        "parameters": [
          {
            "name": "q",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "sort",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "tag",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "folder",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "page",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0
            }
          },
          {
            "name": "per_page",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0
            }
          },
          {
            "name": "deep",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "boolean",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paged bookmark list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad sort mode",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "crate::bookmarks"
        ],
        "operationId": "add_bookmark",
        "requestBody": {
          "description": "Bookmark to add",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Duplicate accepted idempotently",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiBookmark"
                }
              }
            }
          },
          "201": {
            "description": "Bookmark created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiBookmark"
                }
              }
            }
          },
          "400": {
            "description": "Missing url",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Duplicate url",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/bookmarks/{id}": {
      "get": {
        "tags": [
          "crate::bookmarks"
        ],
        "operationId": "get_bookmark",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Short id or UUID prefix",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bookmark",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiBookmark"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "crate::bookmarks"
        ],
        "operationId": "update_bookmark",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Short id or UUID prefix",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Partial fields to update",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated bookmark",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiBookmark"
                }
              }
            }
          },
          "400": {
            "description": "Empty title or url",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "crate::bookmarks"
        ],
        "operationId": "delete_bookmark",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Short id or UUID prefix",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "confirm",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "boolean",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted id, or confirm_required union without confirm",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/bookmarks/{id}/archive": {
      "get": {
        "tags": [
          "crate::bookmarks"
        ],
        "operationId": "get_archive",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Short id or UUID prefix",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Archived HTML page",
            "content": {
              "text/html": {}
            }
          },
          "404": {
            "description": "No archive",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "crate::bookmarks"
        ],
        "operationId": "post_archive",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Short id or UUID prefix",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Backend as {backend?}",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Archived with warnings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "crate::bookmarks"
        ],
        "operationId": "delete_archive",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Short id or UUID prefix",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Archive removed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/bookmarks/{id}/attachments": {
      "post": {
        "tags": [
          "crate::bookmarks"
        ],
        "operationId": "upload_attachment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Short id or UUID prefix",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Base64 attachment as {name, content}",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Attached file name",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad name, payload, or oversize",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/bookmarks/{id}/attachments/{name}": {
      "get": {
        "tags": [
          "crate::bookmarks"
        ],
        "operationId": "download_attachment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Short id or UUID prefix",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "description": "Attachment file name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Attachment bytes",
            "content": {
              "application/octet-stream": {}
            }
          },
          "404": {
            "description": "No such attachment",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "crate::bookmarks"
        ],
        "operationId": "delete_attachment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Short id or UUID prefix",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "description": "Attachment file name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detached file name",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "No such attachment",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/bookmarks/{id}/notes": {
      "get": {
        "tags": [
          "crate::bookmarks"
        ],
        "operationId": "get_notes",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Short id or UUID prefix",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Note body or null",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "crate::bookmarks"
        ],
        "operationId": "put_notes",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Short id or UUID prefix",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Note body as {body}",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Saved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "crate::bookmarks"
        ],
        "operationId": "delete_notes",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Short id or UUID prefix",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notes removed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/bookmarks/{id}/open": {
      "post": {
        "tags": [
          "crate::bookmarks"
        ],
        "operationId": "open_bookmark",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Short id or UUID prefix",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Opened url",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/bulk": {
      "post": {
        "tags": [
          "crate::bulk"
        ],
        "operationId": "bulk",
        "requestBody": {
          "description": "Bulk op as {ids[], op, tags?, folder?, confirm?}",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Confirm gate or deleted/updated counts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "No ids or unknown op",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Id not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/check/apply": {
      "post": {
        "tags": [
          "crate::check"
        ],
        "operationId": "apply_check",
        "requestBody": {
          "description": "Actions as {updates[], quarantine[]}",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated, quarantined, and skipped rows",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/check/run": {
      "post": {
        "tags": [
          "crate::check"
        ],
        "operationId": "run_check",
        "requestBody": {
          "description": "Scope as {spec?, workers?, stale_hours?}",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Checked, ok, and non-ok rows",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad id spec",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/folders/delete": {
      "post": {
        "tags": [
          "crate::taxonomy"
        ],
        "operationId": "delete_folder_ep",
        "requestBody": {
          "description": "Delete as {folder}",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Moved-to-root count",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/folders/rename": {
      "post": {
        "tags": [
          "crate::taxonomy"
        ],
        "operationId": "rename_folder_ep",
        "requestBody": {
          "description": "Rename as {old, new}",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Renamed count",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Same or empty names",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "No bookmarks in that folder",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/history": {
      "get": {
        "tags": [
          "crate::bookmarks"
        ],
        "operationId": "history",
        "parameters": [
          {
            "name": "per_page",
            "in": "query",
            "description": "Page size, clamped 1-500",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number, from 1",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Recently opened bookmarks",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/library/export-site": {
      "post": {
        "tags": [
          "crate::library"
        ],
        "operationId": "export_site_ep",
        "requestBody": {
          "description": "Target as {dir?}",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Written index path",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/library/import": {
      "post": {
        "tags": [
          "crate::library"
        ],
        "operationId": "import_library",
        "requestBody": {
          "description": "Netscape file as {content, markdown?}",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Added, skipped, and warning counts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddRequest": {
        "type": "object",
        "required": [
          "url"
        ],
        "properties": {
          "archive": {
            "type": "boolean"
          },
          "confirm_dup": {
            "type": "boolean"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "folder": {
            "type": [
              "string",
              "null"
            ]
          },
          "markdown": {
            "type": "boolean"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          },
          "url": {
            "type": "string"
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "ListResponse": {
        "type": "object",
        "required": [
          "total",
          "page",
          "per_page",
          "bookmarks"
        ],
        "properties": {
          "bookmarks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiBookmark"
            }
          },
          "page": {
            "type": "integer",
            "minimum": 0
          },
          "per_page": {
            "type": "integer",
            "minimum": 0
          },
          "total": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
    }
  }
}